+++ packages/types/src/create/types.ts
@@ -31,6 +31,7 @@ export enum TypeDefInfo {
 
 export interface TypeDefExtVecFixed {
   length: number;
+  rawName?: string;
   type: string;
 }