+++ packages/types/src/interfaces/identity/types.ts
@@ -3,8 +3,8 @@
 
 import { ITuple } from '@polkadot/types/types';
 import { Enum, Option, Struct, Vec } from '@polkadot/types/codec';
-import { Data, H160, IdentityFields, u32 } from '@polkadot/types/primitive';
-import { AccountId, Balance } from '@polkadot/types/interfaces/runtime';
+import { Data, IdentityFields, u32 } from '@polkadot/types/primitive';
+import { AccountId, Balance, H160 } from '@polkadot/types/interfaces/runtime';
 
 /** @name IdentityInfo */
 export interface IdentityInfo extends Struct {